add `@inline` annotations to `[getindex|setindex!](::ACSet, ...)`
These functions can be significantly optimized with constant information
available within their arguments, so it would be beneficial to enforce
constant propagation for them.
This can be achieved using `Base.@constprop :aggressive` or `@inline`,
and since these functions are likely to be used frequently, I think
promoting inline expansion with `@inline` is a good approach,
so I chose to use `@inline`.

- fixes #121