Declare `Meta` before using it in `@intertypes`
The `@intertypes` macro assigns `mod.Meta` without declaring it.
This was accidentally allowed in Julia 1.9/1.10 due to a dropped
error check, but will likely be disallowed again in Julia 1.11.
See https://github.com/JuliaLang/julia/pull/54678.

This just declares it `global` (which is what happened implicitly
on Julia 1.10/1.11). That said, you may want a `Const` instead,
in which case you would need a different fix.